Generic config-driven template example generator + 8 sparse-topic seeds
- Replace scripts/generate_geo_examples.cjs with the generic
  scripts/generate_template_examples.cjs:
  - Schema: { template_id, params, topics: [...], id_suffix? }
    (legacy `tag: "x"` accepted as topics: [x] so geo_examples_config.json
    keeps working).
  - Files are staged in /tmp/curify_template_examples_<ts>/ first and
    only copied into public/images/ + JSON updated after the batch
    completes — keeps a half-failed run from polluting the repo.
  - Default model: gemini-3-pro-image-preview (Nano Banana Pro).
  - Bilingual generator updated to the same model for consistency.
- New scripts/configs/:
  - 9_traits_stereotypes.json — 5 stereotype variants (Social Butterflies,
    Small Talk Pros, Lone Wolf Thinkers, Born Storytellers, Quiet
    Achievers) for template-9-traits-info-grid.
  - subject_tier3_examples.json — 8 entries seeding the sparsest
    Tier 3 subject topics: nature×2, space×2, weather, body×2, evolution.
- Generated 8 inspirations from subject_tier3_examples.json with Nano
  Banana Pro, copied to public/images/nano_insp{,_preview}/, synced to
  gs://curify-static/.
