+++ public/data/nano_inspiration.json
@@ -42732,5 +42732,179 @@
         "title": "wright brothers vs alberto santos dumont"
       }
     }
+  },
+  {
+    "id": "template-education-nature-coral-reef",
+    "template_id": "template-education",
+    "asset": {
+      "image_url": "/images/nano_insp/template-education-nature-coral-reef.jpg",
+      "preview_image_url": "/images/nano_insp_preview/template-education-nature-coral-reef-prev.jpg"
+    },
+    "params": {
+      "topic": "Coral Reef Ecosystems"
+    },
+    "locales": {
+      "en": {
+        "title": "Coral Reef Ecosystems"
+      }
+    },
+    "topics": [
+      "learning",
+      "cartoon",
+      "nature"
+    ]
+  },
+  {
+    "id": "template-education-card-nature-volcanoes",
+    "template_id": "template-education-card",
+    "asset": {
+      "image_url": "/images/nano_insp/template-education-card-nature-volcanoes.jpg",
+      "preview_image_url": "/images/nano_insp_preview/template-education-card-nature-volcanoes-prev.jpg"
+    },
+    "params": {
+      "topic": "Volcanoes"
+    },
+    "locales": {
+      "en": {
+        "title": "Volcanoes"
+      }
+    },
+    "topics": [
+      "learning",
+      "science",
+      "cartoon",
+      "nature"
+    ]
+  },
+  {
+    "id": "template-education-space-solar-system",
+    "template_id": "template-education",
+    "asset": {
+      "image_url": "/images/nano_insp/template-education-space-solar-system.jpg",
+      "preview_image_url": "/images/nano_insp_preview/template-education-space-solar-system-prev.jpg"
+    },
+    "params": {
+      "topic": "Solar System Planets"
+    },
+    "locales": {
+      "en": {
+        "title": "Solar System Planets"
+      }
+    },
+    "topics": [
+      "learning",
+      "cartoon",
+      "space",
+      "science"
+    ]
+  },
+  {
+    "id": "template-education-card-space-black-holes",
+    "template_id": "template-education-card",
+    "asset": {
+      "image_url": "/images/nano_insp/template-education-card-space-black-holes.jpg",
+      "preview_image_url": "/images/nano_insp_preview/template-education-card-space-black-holes-prev.jpg"
+    },
+    "params": {
+      "topic": "Black Holes Explained"
+    },
+    "locales": {
+      "en": {
+        "title": "Black Holes Explained"
+      }
+    },
+    "topics": [
+      "learning",
+      "science",
+      "cartoon",
+      "space"
+    ]
+  },
+  {
+    "id": "template-education-weather-clouds",
+    "template_id": "template-education",
+    "asset": {
+      "image_url": "/images/nano_insp/template-education-weather-clouds.jpg",
+      "preview_image_url": "/images/nano_insp_preview/template-education-weather-clouds-prev.jpg"
+    },
+    "params": {
+      "topic": "Types of Clouds"
+    },
+    "locales": {
+      "en": {
+        "title": "Types of Clouds"
+      }
+    },
+    "topics": [
+      "learning",
+      "cartoon",
+      "weather"
+    ]
+  },
+  {
+    "id": "template-education-card-body-five-senses",
+    "template_id": "template-education-card",
+    "asset": {
+      "image_url": "/images/nano_insp/template-education-card-body-five-senses.jpg",
+      "preview_image_url": "/images/nano_insp_preview/template-education-card-body-five-senses-prev.jpg"
+    },
+    "params": {
+      "topic": "The Five Senses"
+    },
+    "locales": {
+      "en": {
+        "title": "The Five Senses"
+      }
+    },
+    "topics": [
+      "learning",
+      "science",
+      "cartoon",
+      "body"
+    ]
+  },
+  {
+    "id": "template-education-body-skeletal-system",
+    "template_id": "template-education",
+    "asset": {
+      "image_url": "/images/nano_insp/template-education-body-skeletal-system.jpg",
+      "preview_image_url": "/images/nano_insp_preview/template-education-body-skeletal-system-prev.jpg"
+    },
+    "params": {
+      "topic": "Skeletal System"
+    },
+    "locales": {
+      "en": {
+        "title": "Skeletal System"
+      }
+    },
+    "topics": [
+      "learning",
+      "cartoon",
+      "body",
+      "science"
+    ]
+  },
+  {
+    "id": "template-education-card-evolution-galapagos-finches",
+    "template_id": "template-education-card",
+    "asset": {
+      "image_url": "/images/nano_insp/template-education-card-evolution-galapagos-finches.jpg",
+      "preview_image_url": "/images/nano_insp_preview/template-education-card-evolution-galapagos-finches-prev.jpg"
+    },
+    "params": {
+      "topic": "Galápagos Finches and Natural Selection"
+    },
+    "locales": {
+      "en": {
+        "title": "Galápagos Finches and Natural Selection"
+      }
+    },
+    "topics": [
+      "learning",
+      "science",
+      "cartoon",
+      "evolution"
+    ]
   }
 ]
