+++ scripts/configs/9_traits_stereotypes.json
@@ -1,0 +1,32 @@
+[
+  {
+    "template_id": "template-9-traits-info-grid",
+    "params": { "topic_title": "Social Butterflies" },
+    "topics": ["relationship", "groups", "character"],
+    "id_suffix": "social-butterflies"
+  },
+  {
+    "template_id": "template-9-traits-info-grid",
+    "params": { "topic_title": "Small Talk Pros" },
+    "topics": ["relationship", "expressions", "character"],
+    "id_suffix": "small-talk-pros"
+  },
+  {
+    "template_id": "template-9-traits-info-grid",
+    "params": { "topic_title": "Lone Wolf Thinkers" },
+    "topics": ["character", "personality"],
+    "id_suffix": "lone-wolf-thinkers"
+  },
+  {
+    "template_id": "template-9-traits-info-grid",
+    "params": { "topic_title": "Born Storytellers" },
+    "topics": ["character", "expressions"],
+    "id_suffix": "born-storytellers"
+  },
+  {
+    "template_id": "template-9-traits-info-grid",
+    "params": { "topic_title": "Quiet Achievers" },
+    "topics": ["character", "personality"],
+    "id_suffix": "quiet-achievers"
+  }
+]
